+++ CHANGELOG.md
@@ -1,6 +1,17 @@
+### v1.0.0 2016-03-22
+
+* Add `isTicking` property
+
+* Rename `startClock()` and `stopClock()` to `start()` and `stop()`
+
+* Update to Ember-cli 2.4.2
+
+
+
+
 ### v0.0.2 2015-10-23
 
-* [BUGFIX] Make the `clock-initializer` work with Ember 2.0 and Ember 2.1.
+* [BUGFIX] Make the `clock-initializer` work with Ember 2.0 and Ember 2.1